[ui] show otp in modal
