+++ frontend/src/views/AuthValidate.js
@@ -1,18 +1,24 @@
-import React, { Component } from 'react'
+import React from 'react'
+import { Platform } from 'react-native'
+import { useNavigate } from 'react-router-dom'
+
 import OTPValidate from 'components/Auth/OTPValidate'
 
-import {
-  Box,
-  Heading,
-  View,
-  VStack,
-  useColorMode,
-} from '@gluestack-ui/themed'
-
+import { Box, Button, ButtonText, Heading, VStack } from '@gluestack-ui/themed'
 
 const AuthValidate = (props) => {
+  const navigate = useNavigate()
 
-  const colorMode = useColorMode()
+  const onSuccess = async () => {
+    let url = '/admin/home'
+    try {
+      url = await getAuthReturn()
+    } catch {
+      url = '/admin/home'
+    }
+
+    navigate(url)
+  }
 
   return (
     <Box
@@ -20,11 +26,14 @@ const AuthValidate = (props) => {
       py="$8"
       sx={{
         '@base': { w: '$full', mt: '$1/4' },
-        '@md': { w: '$1/4', mt: '$0', alignSelf: 'center', rounded: 10 }
+        '@md': { w: '$1/4', mt: '$0', alignSelf: 'center', rounded: 10 },
+        _dark: {
+          bg: '$blueGray900'
+        }
       }}
-      bg={colorMode === 'light' ? 'white' : '$blueGray900'}
+      bg="$white"
     >
-      <VStack space="lg">
+      <VStack space="md">
         <Heading
           alignSelf="center"
           size="lg"
@@ -38,11 +47,19 @@ const AuthValidate = (props) => {
         >
           Validate OTP Code
         </Heading>
-        <OTPValidate/>
+
+        <OTPValidate onSuccess={onSuccess} />
+
+        <Button
+          action="secondary"
+          variant="outline"
+          onPress={() => navigate(Platform.OS == 'web' ? -2 : -3)}
+        >
+          <ButtonText>Back</ButtonText>
+        </Button>
       </VStack>
     </Box>
   )
 }
-
 
 export default AuthValidate
